move Feed components
